+++ lab4/src/sum.c
@@ -1,7 +1,7 @@
 #include "sum.h"
 
-int Sum(const struct SumArgs *args) {
-  int sum = 0;
+long long Sum(const struct SumArgs *args) {
+  long long sum = 0;
   for (int i = args->begin; i < args->end; i++) {
     sum += args->array[i];
   }
